# docs/swagger.yaml
basePath: /
definitions:
  handlers.ErrorDetail:
    properties:
      message:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handlers.ErrorDetail'
    type: object
  requests.CreateUserRequest:
    properties:
      email:
        maxLength: 255
        type: string
      username:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - email
    - username
    type: object
  requests.SearchUserRequest:
    properties:
      email:
        maxLength: 255
        type: string
      username:
        maxLength: 255
        minLength: 3
        type: string
    type: object
  responses.ListUsersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.UserResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  responses.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample Go REST API with Swagger documentation.
  title: Golang REST API Boilerplate
  version: "1.0"
paths:
  /v1/users:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ListUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a user
      tags:
      - users
  /v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete a user by ID
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieves a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a user by ID
      tags:
      - users
  /v1/users/search:
    post:
      consumes:
      - application/json
      description: Searches for users
      parameters:
      - description: User to search
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.SearchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Search users
      tags:
      - users
swagger: "2.0"